requirements.txt
# ------------------ Core ------------------
pathspec>=0.12.1
gitpython>=3.1.45
# ------------------ LangChain ------------------
langchain>=0.3.27
langchain-text-splitters>=0.3.11
